Improve get_mutation_level performance
Due to a poorly ordered logic check, literally every time the game queried
whether the player had *any* mutation, it needed to pull their species data
to see if they could bleed (as part of Sanguine Armour's check). Now it
properly verifies that it is looking up Sanguiine Armour *first*.

Additionally, the check to see whether a mutation would be suppressed was
done before even checking to see if the player had the mutation in question
at all. These checks are now reversed.

(Due to a combination of this and the species data map lookup,
get_mutation_level was an unexpected hotspot, requiring a map lookup any
time any mutation was checked - which was often many times per action.)