Improve performance of mons_is_safe
This is called every time a monster_info object is normally created (which
is typically *many* times per turn when around monsters) and involved
several sources of needless effort.

To check whether a monster could reach the player, it set up and performed
full pathfinding for any non-adjacent monster (which allocates multiple
arrays the size of the entire dungeon floor). It now calls can_go_straight
first, as it commonly done in other reachability checks, allowing it to
skip pathfinding in a majority of cases.

It now also skips dispatching to lua in cases where we know that the lua
check will trivially return true. (ch_mon_is_safe can only make monsters
appear *safer*, not more dangerous. So if we've already established that
they're safe, there is no need for this call.

Finally, we don't construct an entire new monster_info to pass to lua when
all the function ever uses is the monster's name (for
runrest_ignore_monster checks).

In combination, this is a moderate measurable improvement to this method's
performance (which can get called quite frequently.)