Add AF_AIRSTRIKE, adjust Airstrike messaging
Crawl's differing elements have very different degrees of representation
throughout the game, in some part influenced each by the different spells
available back in single-digit-version Stone Soup days. Fire gets many
repeats of bolt and fireball and clouds, Ice only rarely gets its
explosion and uses a few gimmicks repeatedly like the Frozen status or
Creeping Frost, Earth gets all sorts of strange connections between metal
and corrosion and slowness and invulnerability, and Air... in terms of
repeatable gimmicks gets Lightning Bolt, repositioning Lightning Bolt, and
Airstrike. The last is at least pretty unique amongst effects in Crawl,
and could help with further flying + air themes for monsters in Crawl, so
expanding it a little further in a weaker context seems reasonable.

This new attack flavour is sort of like casting Airstrike whenever a
monster hits, doing extra damage that scales both with HD and with the
open spaces around the target. It ignores the usual AC check for flavour
triggering (like AF_PURE_FIRE) because it's about manipulating the air
in the space directly around the player (and to make the flat damage
of attacks not be nearly so easily ignored by AC). The brand damage
itself, however, still checks AC, unlike most damage brands. The damage
formula is pretty different from normal Airstrike (in the same way AF_COLD
is nothing like Freeze or Bolt of Cold), with minimal flat damage
component (since that was in the attack itself) and also far more HD and
open space based damage scaling.

      | 0 spaces     | 4 spaces     | 7 spaces
--------------------------------------------------
5 HD  | 4-5 damage   | 9-12 damage  | 14-18 damage
--------------------------------------------------
10 HD | 10-12 damage | 24-28 damage | 35-40 damage
--------------------------------------------------
15 HD | 18-19 damage | 42-46 danage | 61-66 damage

(Strictly speaking, a monster with a polearm _could_ get this brand, but
none of the handful of monster concepts currently planned are going to
wield weapons, so it's extraordinarily unlikely the full 8-space bonus
will ever come up.)

Compared to Airstrike itself, it's inherently less dangerous in terms of
its range and the capacity to dodge or block the initial hit. Of course,
it's also inherently potentially more dangerous if one doesn't actually
respect it enough to move near walls because it'll be activating on
_every_ successful hit instead of just whenever a monster feels like
casting a spell. It should still feel reasonably thematic (since Airstrike
is the iconic one air spell that doesn't use electric damage), and in
particular it may help less observant newcomers notice how Airstrike's
open-air damage bonuses work later on if the brand is given to an
early-game monster. (It's hard to use particularly late in the game in a
number of layouts, but it should be fine in some Lair water branches or
Hells layouts....)

This also comes with some different messages in both the spell and the
attack flavour scaling to the number of open spaces one has around them:
one for 0-2 spaces, one for 3-5 spaces, and one for 6-8 spaces.