Replace current sky beast mechanics with AF_AIRSTRIKE
Sky beasts were added in fe51cf0 for the sheer sake of variety in a very
different development era than the current day. They used to be a source
of mutagenic flesh, but corpse eating has been removed; they used to be
another earlygame monster, but with orc wizards consistently getting
invisibility in the spellbook squashing they're not much of a tutorial
beyond making people start walking backwards for the invisibility to wear
off to instead get random attacks of opportunity; and they use AF_ELEC,
which barely matters compared to more damage considering the lack of
chance to find rElec that early outside of starting off as a Gargoyle.
Still, it is also nice to have early air-themed monsters, as long as Crawl
is going to be reliant in some fashion on the fantasy cliches of the
classical elements, so they deserve a rework rather than removal outright.

As such, they now drop both AF_ELEC and the flickering invisiblity effect
for the new AF_AIRSTRIKE introduced in the previous commit. This swaps out
their 5-6 elec damage (ignoring AC if the first hit deals damage) for a
range between 4-5 physical damage when fully surrounded to a terrifying
14-18 with 7 open air tiles (i.e. with only the sky beast adjacent.) This
seems quite dangerous, but the invisibility bonuses to sky beasts were
quite notable at the point in the game one fought them, and in a full
corridor it'll do about the same damage with slightly more AC-respecting.
It'll kill a lot of blind autofight tactics, but there's not much point
to new monster gimmicks if failing to respect them courts death.

To try to make this new change more obvious, their tile has also been
flipped upside-down. This may be quite silly, but it has always had a
relatively difficult-to-parse tile, and should still visually work as a
ray-like beast swooping down from mid-flight.