Fix a handful of logic errors with wyrmhole hostile teleport
It can no longer teleport the player into no_tele_into locations (though
this may have been unlikely in practice, since it still required LoS to a
monster that *wasn't* standing on a no_tele_into spot.

It no longer picks the first valid spot near a monster, but correctly picks
a random one in LoS of it.

It no longer needlessly counts all monsters in LoS of spots near every
possible monster on the floor (which could cause it to pluralize a message
incorrectly, beyond just wasting time.)

It will fall back on a random teleport if there are no valid spaces near
any valid monster on the floor (instead of not moving the player at all).

It won't perform teleport-related interrupts if you don't actually
teleport.

It won't 'blink' the wyrmhole if the player doesn't actually teleport.

It won't incorrectly amuse Xom even if you don't teleport (for what little
this matters...)

It won't crash if the wyrmhole is dead or gone by the time it actives.

It won't instantly trigger whenever the player presses &B to aim a wizmode
teleport (which even left you with a NON-hostile teleport duration
still active afterward...)

(Also add a few explanatory comments throughout the function.)