Don't leak info about out-of-los monsters via seismosaurus egg (hellmonk)
An exceedingly minor benefit that was probably almost never practical to
do, but the incubating status could indicate the presence of a monster out
of the player's own LoS if it was in the egg's and they were standing next
to it.

Which really doesn't matter very much, but also seems harmless to fix.