Improve aspects of randart property assignment
Currently during randart generation, no artefact property can be
modified if it's intrinsic to the item. This means that artefact rings
of protection from fire can only ever have rF+, for example. While we
don't want to allow such rings to have less fire resistance than rF+,
there should be no problem with them having more.

This commit implements a scheme where an intrinsic property can be
modified, so long as it's in the right direction: it can increase a
property value that's positive or decrease one that's negative, so long
as the appropriate max/min is respected. I've generalized the property
mapping scheme used by jewellery and made a map for talismans to
intrinsic resistance properties by talisman type. Now all items that can
have intrinsic artefact properties are accounted for in a mostly unified
way.

One wrinkle for talismans is the issue of the dragon-blood talisman. We
mark rPois as intrinsic, since all players do get that property in
dragon form, but for the rF++ and rC- there's the issue of draconians
versus non-draconians. It might be reasonable to include those
properties as intrinsic only if the player is non-draconian.