Reroll items that conflict with any fixed artprops
The artprops keyword in an item_spec induces and item to be a random
artefact, but with the specified artefact properties always present. A
problem arises when the base type or sub type of the item is randomized,
since specific base/sub types can conflict with the specified
properties. For example, "any ring artprops:rF:1" will conflict with the
intrinsic rC- of a ring of ice, causing the rF+ artprop to be ignored
when that item type is rolled. For vaults that are trying to define
thematic items, having some randomly generated items ignore the
specified properties won't work. The alternative is to specify an
explicit list of acceptable subtypes in the vault, but this is onerous
and more difficult to maintain.

This commit implements rerolled random types whenever there's a conflict
between the specific fixed artprops and the item's intrinsic properties.
This rerolling is limited to 1000 attempts to avoid an infinite loop, a
scheme already in use before this commit for weapons and jewellery. We
just extend it to staves and talismans and trigger rerolls for all types
whenever there's an artprop conflict.