Remove a special distribution for rings of slaying
Rings of slaying currently have 1/3 the chance to generate as a ring
type compared to other ring types. This was seen as necessary in older
versions, namely prior to 0.10, where slaying was added to a weapon's
base damage and hence was influenced by skills and stats. This is no
longer the case, and rings of slaying aren't markedly more powerful
than, say, a ring of Str/Dex appropriate for the player's weapon type.
This commit removes the special ring of slaying weighting so that all
rings types have the same chance.