Fix Uskayaw linepass confusion targets (klement)
Apparently Usk's linepass confusion has been applying in a full-LOS
beam ahead of the targeted tile for several years and nobody in dev
noticed. Oops? This behavior definitely wasn't intended.

The linepass tracer pierces to determine its targeting, and this is
the same beam that subsequently applies confusion to everything along
its path. This commit changes this beam to use aimed_at_spot.

This is a small power nerf to Usk, which is already in the lower half
of recent god picks, so maybe we'll come up with a minor balance buff
here at some near-future point.

Fixes #4190