Add a hopefully-clarifying statement to Searing Ray cast message (ge0FF)
Searing Ray normally locks onto a monster you cast it on, tracking their
position across several turns. But if you cast the spell with . instead
(ie: the general 'aim at spot' key), it would, well, aim at the spot
instead and *not* track the monster. This was surprising to some players.

I don't feel this behavior is inconsistent with . in general, but it was
said that a message might improve clarity a little, so I have added one to
announce the player's target lock upon casting, if they have one. Possibly
this is a minor improvement?