Tweak Victory placement
Instead of probably wasting a player's superb_item loot half the time,
put Victory on top of the guaranteed down hatch in the central monster
chamber. Since Dig for Victory as popularised in Nethack is supposed to
involve going down a level (in fact multiple levels) to reach Victory,
the hatch is arguably the most appropriate place for this toga!