Replace Deterioration mutation with Poor Constitution
With stat drain being removed, I am adjusting the basic concept of this
mutation to weaken the player in a different way upon taking damage.

This now gives a (reduced, compared to Deterioration) chance to become Weak
for a while upon taking non-trivial damage. At the second level of the
mutation, it gives an additional chance to also be briefly slowed. Numbers
are fairly arbitrary for now.

For the most part, this is still rougher than Deterioration was (as one
level of it rarely did anything relevant to most characters), and it is
possible that some other badmuts could be added to split the weight of it
in future.