Remove GLU dependency
Avoid depending on deprecated GLU library by dynamically loading
OpenGL 3+ function glGenerateMipmap where available.
If it's not available, Where it's not, we just render without mipmapping.

Note that for a lot of resolution / rc setting / zoom level combos,
we don't mipmap anyway, and this PR doesn't change those cases at all.

Also note that OpenGL 3.0 was released in 2008.
glGenerateMipmap is available even on decade old GPUs and iGPUs as a result.
For players running truly ancient GPUs and systems that supported GLU,
but not OpenGL 3, and who happened to have just the right setup for us to
enable mipmapping, we will now render without it.

If anyone actually does turn out to be affected, they may be able to use
a version of llvmpipe that supports OpenGL 3+ as a software renderer.
llvmpipe is available on Linux and Unix-like systems.

Also note we could turn on mipmapping on GLES targets according to
https://registry.khronos.org/OpenGL-Refpages/es3.0/html/glGenerateMipmap.xhtml.

This may be quite useful for low resolution screens, like older android phones.
But I'll leave that for a future commit.