fix: enum related bugfix and cleanups
 * Split explore_stop_options into its own header.
 * Make initfile parsing use maps instead of brute force else-ifs
 * Make the layers enum a enum class
 * simplify base movespeed logic in form_base_movespeed
 * Fix return type of viewmap_flash_clour and fix compiler warning.
 * Fix incorrect enum bitfiddling in viewmap_flash_colour.
Only the last bullet point has any behavioral change.

The intent of a135081b66271c96c04bea8a314a81b1ae5e9791
was to display the flash colour if all 4 layers were
on, or at least that's what the comment said.
The actual behavior was to display it if any of the 4 were on.