Slightly buff the Macabre Finger
In past discussions, it was generally agreed that even if the effect was
cool, this amulet was often a bit underwhelming in practice, so let's give
it a little extra boost. (The idea of making it a regeneration amulet had
been proposed, but that might make it *too* automatically good. This should
be something like '1/2 regen' in a lot of cases, but texturally different.)

Was the number 7 really lucky for whatever poor soul this finger once was?

(Also, change the artprop name to match Vainglory)