Fix a couple crashes related to equipping dangerous items on the floor
If one attempted to equip an item on the floor such that a warning prompt
would happen (ie: for ^Drain items), it could cause a crash in two
different ways - by trying to get an out of bounds inventory letter, or by
leaving a 'ghost' item in the player's backpack that could later cause the
game to think the player's inventory had room, but fail to put anything in
it.

But I think the entire process of temporarily putting the item into the
player's inventory is no longer necessary (it only existed due to stat zero
checks from an older version that never even made it to trunk), so that
alone should hopefully fix these crashes.