Update the commercial district of grated_community_mu
The Legendary Smithy shop in the grated_community vault was not a
particularly good shop. A shop doesn't have to always be useful, but
it should at least be useful sometimes. A shop that's always crap is
a joke that's only ever funny once. Any possible usefulness for
mundane, damaged high-end gear is a gimmick done better by the
Lightly-Used Armoury anyway. Accordingly, use gammafunk's new
artprops specification tech to give it a new gimmick, similar to the
old one, but with the potential to make something someone might
conceivably ever use: All randarts, but each has at least one bad
artefact property.

Also, replace the long-lost food shop with a scroll shop. Similarly
to the Miraculous Elixirs next door, it sells only scrolls of
non-guaranteed utility: noise, torment, and vulnerability. It's not
quite the same as a potion of mutation, but it'll do for now.