Remove ac_with_specific_items functions
These were once used as part of the early AC change previews, but the
current system that actually puts on arbitrary items temporarily has
superceded that for a while. The only remaining user of these functions for
anything other than 'the armour you are actually wearing' is the
description for !lig, and that can use the same logic as forms checking
for flight (ie: temporarily transform and then look at our resulting
stats). This is a lot heavier weight, but for this purpose that's fine (and
by the code's own admission, it didn't properly restrict itself to the
passed set of items anyway.

This also fixes shields of protection no longer giving AC.