Fix some weird effects with Ash and the tiles item window (Darby)
If someone was worshipping Ash, had some cursed equipment, and moved their
mouse over equipment in the local tiles inventory panel, it could cause
the visible equip status of those items to change (and also print messages
to the log and maybe even identify items....)

Basically, mouseover creates the item description tooltip, which does the
equipment preview, which quietly swaps equipment in the background to see
your new stats. But because one call to ash_check_bondage() was in the
wrong place, doing so could change your piety in the midle of the preview,
causing messages to print and redrawing the inventory in the middle of
what should otherwise be an invisible action.

This is hopefully fixed now.