New trap: tyrant's trap, weakening players + mighting monsters
With the steady removal of mechanical / direct damage traps, there have
been strange consequences as well as large amounts of design space both
opened up by the lack of weaker trap types. The orb chamber's double-trap
arrangement can lead to fiddly, tedious tactics at what is meant to be the
climax of the game, Tomb has continued to creep more steadily upwards in
threat due to its extreme reliance on torment, and vaults can't establish
a place as fortified without jumping straight to alarm and net traps at
minimum (incidentally losing most of the original Ossuary gimmick).

As such, I'm starting up a set of weaker traps, aimed both at helping out
these cases and providing more flexible + tangible flavour for both vaults
and given branches than The Power of Zot or translocations or infinite
net machines. They will still follow the same current paradigm as Zot and
alarm traps, where players and hostiles alike walking onto a trap hurls an
effect at the player, to help keep the nature of traps straightforward
enough to easily learn.

Tyrant's traps are for militaristic contexts. Whenever the player or a
hostile monster steps on the trap, it will steal from the player's melee
capabilities (by inflicting 10-14 turns of Weak, capped at 50) in order to
empower two or three highest-hd monsters in sight (by granting 20-28 turns
of Might). If there are no monsters in sight, the Weak is still applied.
(While this and other upcoming traps could mildly annoying to rest off if
there aren't any further monsters in sight, with pre-revealed traps it's
rare to need to activate traps to explore a floor, and when it does block
off a path it's nowhere near the annoyance dispersal and teleport traps
can invoke in the same capacity, so it should be a net reduction in
annoyance.) This swing should hopefully force melee-centric characters
to switch tactics, while still being relevant for other characters as
Might has shown to work throughout the game.

These traps rarely spawn naturally in Orc, Snake, and Vaults. Orc could
do with more unique components beyond the orcs regularly used elsewhere,
Snake struggles in terms of variety and distinction versus the other Lair
branches (plus it contains both status effects already), and V needs
more thematic aspects to help establish it as an organized branch
versus the Depths shortly after it. Each of these traps will come with a
slight decrease for shafts (which are now mostly helpful for players in
the actually visible form) and dispersal traps (which are roughly
comparable in helping monsters and hurting the player on a short-term
basis).

(The messaging is kind of weird- I'm not super happy with my options, but
"constitution" reads as max hp drain and "power" is used regularly for mp.
With no actual statdrain in the game anymore as of de20453, hopefully
people will notice the Weak status being gained instead of expecting
actual lost "strength".)

The tile was composited from CC0 level block scraps by Surt and Bloax's
original tentacled monstrosity tile.