Fix religious abilities having the wrong hotkeys for zealot starts (Kab)
God invocations default to being mapped to a/b/c/etc. and Renounce Religion
is mapped to 'X'. But since 2dbd7c1 zealot starts would have Renounce
Religion on 'f', and god abilities like berserk start at 'g'. (Joining said
god later on worked properly.)

The reason for this was rather opaque, but essentially: the new equip
system causes equip_item() to be called for the player's starting equipment
and this calls check_item_hint() which grabs a list of the player's
current abilities (to see if they have gained any, so that hints mode can
teach the player about that). Getting a list of abilities automatically
applies hotkeys to any the player has which don't have any yet (and the
new character already worships their god at this point, so Renounce
Religion, Berserk, etc. are given default keybinds.) By the time
set_god_ability_slots() was called, they already had keybinds and so were
not given any, resulting in this issue.

I have elected to resolve this by having set_god_ability_slots() delete
even abilities that belong to your current god, so they can be assigned
fresh (this is only ever called in newgame and upon joining a god, so this
seems safe.). Renounce Religion is handled slightly differently and given
a default keybind in find_ability_slot(). This technically means that if
the player has something already assigned to X when they join a god, that
it will end up elsewhere instead. I'm not sure this matters.